double change

+++ calculadoraSwitch/src/calculadoraSwitch/Division.java
@@ -6,24 +6,24 @@ public class Division {
 	
 	Scanner dato = new Scanner(System.in);
 	
-	int tamano, div = 0, temp = 0;
+	double tamano, div = 0, temp = 0;
 	public void operacionDiv() {
 		
 	System.out.println("¿Cuantos valores desea dividir?");
 	tamano = dato.nextInt();
-	int numero[] = new int[tamano];
+	double numero[] = new double[(int) tamano];
 
-	for(int i = 0; i < tamano; i++){
+	for(double i = 0; i < tamano; i++){
 	    System.out.println(" Valor # " + i + "\nIngrese numero");
-	    numero[i] = dato.nextInt();
+	    numero[(int) i] = dato.nextDouble();
 	    if(i == (tamano - 1)){
-	        temp = numero[i]; 
+	        temp = numero[(int) i]; 
 	    }
 	} 
 
 	
-	for(int x = (tamano - 2); x >= 0; x--) { 
-		div =  numero[x] / temp; 
+	for(double x = (tamano - 2); x >= 0; x--) { 
+		div =  numero[(int) x] / temp; 
 	    temp = div; 
 	}
 	System.out.println("La division es: "+ div);
